requirements.txt
tensorflow==2.10.0
fastapi
translate
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib